Generalize exchange to complex orbitals D1/D2
Currently exchange is on-the-fly only, i.e. no
setupInternal() is used. Exchange now requires
bank whenever MPI > 1. The D2 exchange is currently
NOT correct when MPI > 1.

Tweak exchange_prec input:
Now a negative value means it will follow the
setup precision, which is the current precision
used in the SCF cycle.